+++ src/main/java/com/academy/mudogroupware/workspace/domain/exception/WorkspaceLastMemberException.java
@@ -1,0 +1,10 @@
+package com.academy.mudogroupware.workspace.domain.exception;
+
+import com.academy.mudogroupware.global.domain.common.exception.BadRequestException;
+
+public class WorkspaceLastMemberException extends BadRequestException {
+
+  public WorkspaceLastMemberException() {
+    super(WorkspaceErrorCode.LAST_MEMBER_CANNOT_LEAVE);
+  }
+}